Separate message handlers to content and metadata
